[cleanup] Moved updateAllPinocchio function to pinocchio factory
