+++ unittest/factory/pinocchio_model.hpp
@@ -6,8 +6,17 @@
 // All rights reserved.
 ///////////////////////////////////////////////////////////////////////////////
 
+#include <pinocchio/fwd.hpp>
+#include <pinocchio/algorithm/frames.hpp>
+#include <pinocchio/algorithm/jacobian.hpp>
+#include <pinocchio/algorithm/kinematics.hpp>
+#include <pinocchio/algorithm/center-of-mass.hpp>
+#include <pinocchio/algorithm/centroidal.hpp>
+#include <pinocchio/algorithm/kinematics-derivatives.hpp>
+#include <pinocchio/algorithm/centroidal-derivatives.hpp>
 #include <pinocchio/parsers/urdf.hpp>
 #include <pinocchio/parsers/sample-models.hpp>
+
 #include <example-robot-data/path.hpp>
 
 #ifndef CROCODDYL_PINOCCHIO_MODEL_FACTORY_HPP_
@@ -112,6 +121,30 @@ class PinocchioModelFactory {
   std::string frame_name_;                     //!< Frame name for unittesting
   std::size_t frame_id_;                       //!< Frame id for unittesting
 };
+
+/**
+ * @brief Compute all the pinocchio data needed for the numerical
+ * differentiation. We use the address of the object to avoid a copy from the
+ * "boost::bind".
+ *
+ * @param model is the rigid body robot model.
+ * @param data contains the results of the computations.
+ * @param x is the state vector.
+ */
+void updateAllPinocchio(pinocchio::Model* const model, pinocchio::Data* data, const Eigen::VectorXd& x) {
+  const Eigen::VectorXd& q = x.segment(0, model->nq);
+  const Eigen::VectorXd& v = x.segment(model->nq, model->nv);
+  Eigen::VectorXd a = Eigen::VectorXd::Zero(model->nv);
+  Eigen::Matrix<double, 6, Eigen::Dynamic> tmp;
+  tmp.resize(6, model->nv);
+  pinocchio::forwardKinematics(*model, *data, q);
+  pinocchio::computeForwardKinematicsDerivatives(*model, *data, q, v, a);
+  pinocchio::computeJointJacobians(*model, *data, q);
+  pinocchio::updateFramePlacements(*model, *data);
+  pinocchio::jacobianCenterOfMass(*model, *data, q);
+  pinocchio::computeCentroidalMomentum(*model, *data, q, v);
+  pinocchio::computeCentroidalDynamicsDerivatives(*model, *data, q, v, a, tmp, tmp, tmp, tmp);
+}
 
 }  // namespace crocoddyl_unit_test
